move Bundle related methods
