+++ navigator/androidx-nav/api/androidx-nav.api
@@ -1,3 +1,6 @@
+public final class com/freeletics/mad/navigator/internal/BundleKt {
+}
+
 public final class com/freeletics/mad/navigator/internal/DestinationIdKt {
 }
