+++ navigator/runtime/src/main/java/com/freeletics/mad/navigator/internal/Navigate.kt
@@ -1,11 +1,7 @@
 package com.freeletics.mad.navigator.internal
 
-import android.os.Bundle
-import android.os.Parcelable
 import androidx.activity.result.ActivityResultLauncher
 import com.freeletics.mad.navigator.ActivityResultRequest
-import com.freeletics.mad.navigator.ActivityRoute
-import com.freeletics.mad.navigator.BaseRoute
 import com.freeletics.mad.navigator.NavEvent
 import com.freeletics.mad.navigator.PermissionsResultRequest
 
@@ -58,33 +54,3 @@ public fun navigate(
         }
     }
 }
-
-
-@InternalNavigatorApi
-public fun <T : BaseRoute> Bundle?.requireRoute(): T = requireNotNull(this) {
-        "Bundle is null. Can't extract Route data."
-    }
-    .run {
-        requireNotNull(getParcelable(EXTRA_ROUTE)) {
-            "Bundle doesn't contain Route data in \"$EXTRA_ROUTE\""
-        }
-    }
-
-@InternalNavigatorApi
-public fun BaseRoute.getArguments(): Bundle = Bundle().also {
-    it.putParcelable(EXTRA_ROUTE, this)
-}
-
-@InternalNavigatorApi
-public fun ActivityRoute.getArguments(): Bundle = Bundle().also {
-    it.putParcelable(EXTRA_FILL_IN_INTENT, fillInIntent())
-    if (this is Parcelable) {
-        it.putParcelable(EXTRA_ROUTE, this)
-    }
-}
-
-@InternalNavigatorApi
-public const val EXTRA_ROUTE: String = "com.freeletics.mad.navigation.ROUTE"
-
-@InternalNavigatorApi
-public const val EXTRA_FILL_IN_INTENT: String = "com.freeletics.mad.navigation.FILL_IN_INTENT"
